requirements.txt
fastapi
uvicorn
yfinance
pandas
python-dotenv
pydantic
langchain
langchain-openai
openai
pytest